feat: 竞赛题目的编辑和竞赛的发布和取消发布

+++ oj-modules/oj-system/src/main/java/com/example/system/TimeCheck/inte/NotLaterThan.java
@@ -16,7 +16,7 @@ public @interface NotLaterThan {
     String endTimeField();
 
     // 新增：最小小时差，默认2小时
-    int minHours() default 2;
+    long minHours() default 2;
 
     Class<?>[] groups() default {};
     Class<? extends Payload>[] payload() default {};
